refactor: split the memory broker TestClient into its own module
src/memory/mod.rs sat at 537 lines, over the ~500-line guideline. The
`impl TestClient for MemoryBroker` block is a self-contained concern (the
broker doubling as its own test client) and the only natural seam.

Move it into a child `memory::test_client` module. A child module sees its
parent's private items, so the impl still reaches the private `MemoryState`
fields (`published`, `notify`) and `MemoryBroker::state` with no visibility
changes. mod.rs drops to 463 lines; the `timeout` import and the now-unused
`testing::TestClient` import move with the impl.

Pure move, no behaviour change. Every public path stays byte-identical.
